pyproject.toml
[project]
name = "langraph-ecommerce-agent"
version = "0.1.0"
description = "custom agent using the LangGraph framework that processes e-commerce data and derives meaningful business insights"
requires-python = ">=3.13"
dependencies = [
    "db-dtypes>=1.4.4",
    "fastapi>=0.123.0",
    "google-cloud-bigquery>=3.38.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "langchain>=1.1.0",
    "langgraph>=1.0.4",
    "orjson>=3.11.4",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.38.0",
]


[dependency-groups]
dev = ["ipykernel>=6.30.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/analyst_9000"]


#[project.scripts]
#smith = "xgs_smith.main:run_app_locally"
